pyproject.toml
[project]
name = "PersonalPortfolioManagerAI"
version = "0.1.0"
description = "UV is better than poetry"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.3.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "newspaper3k>=0.2.8"
]
